Student-Projects/group22/requirements.txt
pytesseract==0.3.10
pillow==10.3.0
python-telegram-bot==21.5